requirements.txt
Django==1.6
argparse==1.2.1
distribute==0.6.24
django-zurb-foundation==5.0.2
wsgiref==0.1.2
django-tagging
django-taggit
